FileHistoryMenus is a nested class of ProjectManager
