+++ src/ProjectManager.cpp
@@ -50,6 +50,7 @@ Paul Licameli split from AudacityProject.cpp
 #include "widgets/WindowAccessible.h"
 
 #include <wx/app.h>
+#include <wx/menu.h>
 #include <wx/scrolbar.h>
 #include <wx/sizer.h>
 #include <wx/splitter.h>
@@ -887,4 +888,78 @@ int ProjectManager::GetEstimatedRecordingMinsLeftOnDisk(long lCaptureChannels) {
    // Convert to minutes before returning
    int iRecMins = (int)round(dRecTime / 60.0);
    return iRecMins;
+}
+
+void ProjectManager::UseMenu(wxMenu *menu)
+{
+   FileHistoryMenus::Instance().UseMenu(menu);
+}
+
+void ProjectManager::FileHistoryMenus::UseMenu(wxMenu *menu)
+{
+   Compress();
+
+   auto end = mMenus.end();
+   auto iter = std::find(mMenus.begin(), end, menu);
+   auto found = (iter != end);
+
+   if (!found)
+      mMenus.push_back(menu);
+   else {
+      wxASSERT(false);
+   }
+
+   NotifyMenu( menu );
+}
+
+ProjectManager::FileHistoryMenus::FileHistoryMenus()
+{
+   mSubscription = FileHistory::Global()
+      .Subscribe(*this, &FileHistoryMenus::OnChangedHistory);
+}
+
+ProjectManager::FileHistoryMenus &ProjectManager::FileHistoryMenus::Instance()
+{
+   static FileHistoryMenus instance;
+   return instance;
+}
+
+void ProjectManager::FileHistoryMenus::OnChangedHistory(Observer::Message)
+{
+   Compress();
+   for (auto pMenu : mMenus)
+      if (pMenu)
+         NotifyMenu(pMenu);
+}
+
+void ProjectManager::FileHistoryMenus::NotifyMenu(wxMenu *menu)
+{
+   wxMenuItemList items = menu->GetMenuItems();
+   for (auto end = items.end(), iter = items.begin(); iter != end;)
+      menu->Destroy(*iter++);
+
+   const auto &history = FileHistory::Global();
+   int mIDBase = ID_RECENT_CLEAR;
+   int i = 0;
+   for (auto item : history) {
+      item.Replace( "&", "&&" );
+      menu->Append(mIDBase + 1 + i++, item);
+   }
+
+   if (history.size() > 0) {
+      menu->AppendSeparator();
+   }
+   menu->Append(mIDBase, _("&Clear"));
+   menu->Enable(mIDBase, history.size() > 0);
+}
+
+void ProjectManager::FileHistoryMenus::Compress()
+{
+   // Clear up expired weak pointers
+   auto end = mMenus.end();
+   mMenus.erase(
+     std::remove_if( mMenus.begin(), end,
+        [](wxWeakRef<wxMenu> &pMenu){ return !pMenu; } ),
+     end
+   );
 }
